lmapd: avoid some int/uint comparisons
Change some auxiliary/loop variable types to unsigned types, to avoid
sign conversions and comparisons between signed and unsigned types.

This might result in slightly worse machine code, bit it would still be
a worthy tradeoff.

CEPTRO.br-issue: MDC-1521

+++ src/runner.c
@@ -83,7 +83,7 @@ safe_event_free(struct event * event)
 static unsigned int
 rand_interval(unsigned int min, unsigned int max)
 {
-    int r;
+    unsigned int r;
     const unsigned int range = 1 + max - min;
     const unsigned int buckets = RAND_MAX / range;
     const unsigned int limit = buckets * range;
@@ -92,7 +92,7 @@ rand_interval(unsigned int min, unsigned int max)
      * the buckets until you land in one of them. All buckets are equally
      * likely. If you land off the end of the line of buckets, try again. */
     do {
-	r = rand();
+	r = (unsigned int)rand();
     } while (r >= limit);
 
     return min + (r / buckets);
